move and rename everything
