+++ examples/.interop/next-minimal-starter-migration/src/utils/trpc.ts
@@ -1,7 +1,7 @@
-import { createReactQueryHooks, createReactQueryHooksProxy } from '@trpc/react';
+import { createReactQueryHooks } from '@trpc/react';
 import type { AppRouter } from '../pages/api/trpc/[trpc]';
 
 export const trpc = createReactQueryHooks<AppRouter>();
 // => { useQuery: ..., useMutation: ...}
 
-export const proxy = createReactQueryHooksProxy(trpc);
+export const proxy = trpc.proxy;
