+++ packages/react/src/index.ts
@@ -1,14 +1,8 @@
 export * from '@trpc/client';
 
-export {
-  createReactQueryHooks,
-  type CreateReactQueryHooks,
-} from './createReactQueryHooks';
-export {
-  createReactQueryHooksProxy,
-  createTRPCReact,
-} from './createReactQueryHooksProxy';
+export { createTRPCReact } from './createTRPCReact';
 export type {
   DecoratedProcedureUtilsRecord,
   DecoratedProcedureRecord,
 } from './shared';
+export { createReactQueryHooks } from './interop';
