+++ packages/react/src/shared/index.ts
@@ -1,3 +1,3 @@
 export * from './proxy/decorationProxy';
 export * from './proxy/utilsProxy';
-export type { DecoratedProcedureRecord } from '../createReactQueryHooksProxy';
+export type { DecoratedProcedureRecord } from '../createTRPCReact';
